lab_connector/requirements.txt
httpx>=0.27,<1
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
python-dotenv>=1.0,<2
